Add an `include` flag to only include certain messages
By adding this option, editors have a way to run `gometalinter` and
only output the results for the file being saved. Of course it also
helps other use cases where you only want to check if a specific
error/warning is triggered by your code.

Tested this with a customised version of `vim-go` (will make a PR there
as well to use this new functionality) and it works really well.