Change {path} to {paths} in README example
I assume this is a typo in the README. A basic test shows that `{path}` is the correct keyword, whereas `{paths}` doesn't do anything special. I tested this via the following command:

    gometalinter --vendor --disable-all --linter='vet:bash -c "echo {path} >~/Desktop/proof":PATH:LINE:MESSAGE' --enable=vet .

The above command writes the  `.` (or whatever path you choose to pass to the linter) to the file. Replacing `{path}` with `{paths}` causes "{paths}" to be written to the file, which is clearly not what we want.