misspell: Add locale setting to misspell
This adds a --misspell-locale option to gometalinter. It passes the
desired locale setting to misspell so that you can check for common
typos like "color" which are obviously supposed to read "colour". ;)
Unfortunately there is no Canadian linter at this time, so UK will have
to do for now.

Example:

gometalinter --disable-all --enable=misspell --misspell-locale=UK