Install only enabled linters
The introduces the idea of only installing enabled linters. This
is useful when you know that you're only going to ever run a certain
linter. It reduces the number of potential moving parts along with
the added advantage of hastening the install process. This especially
useful for a CI workflow.

Running the following shows the changes:

```
gometalinter --disable-all --enable=goimports --install
```