Fix: gometalinter can't work on multi-GOPATH
The function call packageNameFromPath(path) will return a wrong result
when `path` is not in the first GOPATH. It just returns the `relpath` which
relative to the first GOPATH, but the source code is not in there in fact.

This commit will try next GOPATH when got a wrong `relpath`,
and try os.Getwd() when all GOPATH failed.

This should close issue #478.